dapper repositories

+++ EducationApp.DataAccessLayer/Models/GenericModel.cs
@@ -7,6 +7,6 @@ namespace EducationApp.DataAccessLayer.Models
     public class GenericModel<T> where T : class
     {
         public IEnumerable<T> Collection = new List<T>();
-        public long CollectionCount { get; set; }
+        public int CollectionCount { get; set; }
     }
 }
